% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition.R
\name{build_nutrition_table_html}
\alias{build_nutrition_table_html}
\title{Build HTML nutrition table for a recipe.}
\usage{
build_nutrition_table_html(nutrition_result, portions = NA_real_)
}
\arguments{
\item{nutrition_result}{Output of \code{\link[=calc_recipe_nutrition]{calc_recipe_nutrition()}}.}

\item{portions}{Numeric portion count.}
}
\value{
A single HTML string.
}
\description{
Build HTML nutrition table for a recipe.
}
