% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qmds.R
\name{gha_generate_qmds}
\alias{gha_generate_qmds}
\title{GitHub Actions entrypoint: generate QMD files from recipe YAMLs.}
\usage{
gha_generate_qmds(recipes_dir = "recettes", images_dir = "images")
}
\description{
GitHub Actions entrypoint: generate QMD files from recipe YAMLs.
}
