% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_import_url.R
\name{gha_import_recipe_from_url}
\alias{gha_import_recipe_from_url}
\title{GitHub Actions entrypoint: import a recipe from a generic URL YAML request.}
\usage{
gha_import_recipe_from_url(url_file = Sys.getenv("RECIPE_URL_FILE"))
}
\description{
GitHub Actions entrypoint: import a recipe from a generic URL YAML request.
}
