% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_import_youtube.R
\name{gha_import_recipe_from_youtube}
\alias{gha_import_recipe_from_youtube}
\title{GitHub Actions entrypoint: import a recipe from a YouTube URL YAML request.}
\usage{
gha_import_recipe_from_youtube(url_file = Sys.getenv("RECIPE_URL_FILE"))
}
\description{
GitHub Actions entrypoint: import a recipe from a YouTube URL YAML request.
}
