% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{gha_validate_recipes}
\alias{gha_validate_recipes}
\title{GitHub Actions entrypoint: validate recipe YAML files.}
\usage{
gha_validate_recipes(recipes_dir = "recettes")
}
\description{
GitHub Actions entrypoint: validate recipe YAML files.
}
