% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livrec-package.R
\docType{package}
\name{livrec-package}
\alias{livrec}
\alias{livrec-package}
\title{livrec: Recipe automation utilities}
\description{
Internal package used by the recettes project to validate recipes, import
recipes from URLs/YouTube, compute nutrition, and generate Quarto pages.
}
\author{
\strong{Maintainer}: Alexandre Parent \email{alexandre@example.com}

}
\keyword{internal}
