% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_qmd.R
\name{regenerate_recipe_qmds}
\alias{regenerate_recipe_qmds}
\title{Regenerate all recipe QMDs from YAMLs}
\usage{
regenerate_recipe_qmds(recipes_dir = "recettes", pattern = "\\\\.ya?ml$")
}
\arguments{
\item{recipes_dir}{Directory containing YAML recipes.}

\item{pattern}{File matching pattern for YAML recipes.}
}
\value{
Vector of processed YAML file paths invisibly.
}
\description{
Regenerate all recipe QMDs from YAMLs
}
