% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_qmd.R
\name{yaml_recipe_to_qmd}
\alias{yaml_recipe_to_qmd}
\title{Convert one recipe YAML file to Quarto (\code{.qmd}).}
\usage{
yaml_recipe_to_qmd(yaml_path, qmd_path = NULL)
}
\arguments{
\item{yaml_path}{Path to source YAML recipe file.}

\item{qmd_path}{Optional output \code{.qmd} path. Defaults to same path with
extension replaced by \code{.qmd}.}
}
\value{
Output \code{.qmd} path invisibly.
}
\description{
Convert one recipe YAML file to Quarto (\code{.qmd}).
}
