% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnf.R
\name{gha_build_cnf_foods}
\alias{gha_build_cnf_foods}
\title{GitHub Actions entrypoint: build CNF-based nutrition file.}
\usage{
gha_build_cnf_foods(
  cnf_dir = Sys.getenv("CNF_DIR", unset = "data/nutrition/cnf"),
  map_path = Sys.getenv("CNF_MAP_PATH", unset = "data/nutrition/cnf_food_map.csv"),
  base_path = Sys.getenv("BASE_FOODS_PATH", unset = "data/nutrition/foods.csv"),
  out_path = Sys.getenv("CNF_OUT_PATH", unset = "data/nutrition/foods_cnf.csv")
)
}
\description{
GitHub Actions entrypoint: build CNF-based nutrition file.
}
