% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_to_qmd.R
\name{regenerate_recipe_qmds}
\alias{regenerate_recipe_qmds}
\title{Regenerate all recipe QMDs from YAMLs}
\usage{
regenerate_recipe_qmds(recipes_dir = "recettes", pattern = "\\\\.ya?ml$")
}
\description{
Regenerate all recipe QMDs from YAMLs
}
